fix: move all mocks to setupTests.js for global application
- Move fetch, SockJS, and STOMP Client mocks to setupTests.js
- This ensures mocks are applied before any module imports
- Fixes 'Cannot read properties of undefined' and 'activate is not a function' errors
- All mocks now apply globally to all test files
- Simplify test files by removing duplicate mock definitions
